requirements.txt
google-genai
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dateutil